Update "chemist" alias
Background by thiemowmde:

"chemist" is heavily ambiguous. Could be a science lab on a
campus. Could be an industrial site. Could be an alternative for a
pharmacy.  A quick image search leaves the impression that no country
uses a test glass to represent a pharmacy. The Maki 2 announcement
mentions Noun Project as inspiration. Searching that for "chemist"
and "pharmacy" doesn't make the situation much clearer.  The patch
that added the icon associates it with many different things.
dewikivoyage translates it as "Drogerie".  A Maki 2 issue from 2016
associates it with "pharmacy".

There are many possible replacements for "minefield", e.g.  "barrier"
or "caution". "danger" felt most appropriate.

"london-underground", "rail-above" and "rail-underground" become
"rail". I couldn't find a better alternative.